chore: Rewrite page and redirects processor
