+++ src/pages/index.ts
@@ -1,2 +1,1 @@
 export {generateStaticMarkup} from './document';
-export {generateStaticRedirect} from './redirect';
