+++ src/steps/index.ts
@@ -1,2 +1,1 @@
-export * from './processPages';
 export * from './processChangelogs';
