+++ src/utils/index.ts
@@ -1,2 +1,10 @@
-export * from './singlePage';
-export * from './path';
+export function getDepth(path: string) {
+    return path
+        .replace(/\\/g, '/')
+        .replace(/^\.\/|\/$/g, '')
+        .split('/').length;
+}
+
+export function getDepthPath(depth: number) {
+    return Array(depth).fill('../').join('') || './';
+}
