--- src/utils/path.ts
@@ -1,10 +1,0 @@
-export function getDepth(path: string) {
-    return path
-        .replace(/\\/g, '/')
-        .replace(/^\.\/|\/$/g, '')
-        .split('/').length;
-}
-
-export function getDepthPath(depth: number) {
-    return Array(depth).fill('../').join('') || './';
-}
